# backend/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "CreateOrganizationRequest": {
        "properties": {
          "displayName": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "publicContact": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "slug": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "slug"
        ],
        "type": "object"
      },
...
      "OrganizationView": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "publicContact": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MiezMerker HTTP API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "listMyOrganizations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Organizations of the authenticated user (via ACTIVE or any membership)",
        "tags": [
          "organization-controller"
        ]
      },
      "post": {
        "operationId": "createOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create an organization; caller becomes ACTIVE ADMIN (test-phase provisioning)",
        "tags": [
          "organization-controller"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}": {
      "get": {
        "operationId": "getOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Organization details; requires ACTIVE membership (tenant isolation)",
        "tags": [
          "organization-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}